Extract Extensions into their own right panel tab
