+++ src/components/views/right_panel/RightPanelTabs.tsx
@@ -16,6 +16,7 @@ limitations under the License.
 
 import React, { useRef } from "react";
 import { NavBar, NavItem } from "@vector-im/compound-web";
+import { Room } from "matrix-js-sdk/src/matrix";
 
 import { _t } from "../../../languageHandler";
 import { RightPanelPhases } from "../../../stores/right-panel/RightPanelStorePhases";
@@ -24,17 +25,27 @@ import PosthogTrackers from "../../../PosthogTrackers";
 import { useDispatcher } from "../../../hooks/useDispatcher";
 import dispatcher from "../../../dispatcher/dispatcher";
 import { Action } from "../../../dispatcher/actions";
+import SettingsStore from "../../../settings/SettingsStore";
+import { UIComponent, UIFeature } from "../../../settings/UIFeature";
+import { shouldShowComponent } from "../../../customisations/helpers/UIComponents";
+import { useIsVideoRoom } from "../../../utils/video-rooms";
 
 function shouldShowTabsForPhase(phase?: RightPanelPhases): boolean {
-    const tabs = [RightPanelPhases.RoomSummary, RightPanelPhases.RoomMemberList, RightPanelPhases.ThreadPanel];
+    const tabs = [
+        RightPanelPhases.RoomSummary,
+        RightPanelPhases.RoomMemberList,
+        RightPanelPhases.ThreadPanel,
+        RightPanelPhases.Extensions,
+    ];
     return !!phase && tabs.includes(phase);
 }
 
 type Props = {
+    room?: Room;
     phase: RightPanelPhases;
 };
 
-export const RightPanelTabs: React.FC<Props> = ({ phase }): JSX.Element | null => {
+export const RightPanelTabs: React.FC<Props> = ({ phase, room }): JSX.Element | null => {
     const threadsTabRef = useRef<HTMLButtonElement | null>(null);
 
     useDispatcher(dispatcher, (payload) => {
@@ -44,6 +55,8 @@ export const RightPanelTabs: React.FC<Props> = ({ phase }): JSX.Element | null =
             threadsTabRef.current?.focus();
         }
     });
+
+    const isVideoRoom = useIsVideoRoom(room);
 
     if (!shouldShowTabsForPhase(phase)) return null;
 
@@ -81,6 +94,20 @@ export const RightPanelTabs: React.FC<Props> = ({ phase }): JSX.Element | null =
             >
                 {_t("common|threads")}
             </NavItem>
+            {SettingsStore.getValue(UIFeature.Widgets) &&
+                !isVideoRoom &&
+                shouldShowComponent(UIComponent.AddIntegrations) && (
+                    <NavItem
+                        aria-controls="thread-panel"
+                        id="extensions-panel-tab"
+                        onClick={() => {
+                            RightPanelStore.instance.pushCard({ phase: RightPanelPhases.Extensions }, true);
+                        }}
+                        active={phase === RightPanelPhases.Extensions}
+                    >
+                        {_t("common|extensions")}
+                    </NavItem>
+                )}
         </NavBar>
     );
 };
